Support using named parameters
In various places, rest parameter syntax (e.g. `...params') was used to
represent the parameters to an SQL statement. The problem with this
approach is that it would not allow named parameters to be used, because
in order to use named parameters, an object must be passed as the params
to the driver method, but passing an object as the first parameter would
end up being wrapped in an array, and the driver method would treat the
object as a single parameter.

This commit replaces the uses of rest parameter syntax with a check
whether there is exactly one parameter whose type is 'object'. If so,
that object is used as the params to the driver method. Otherwise, an
appropriate slice of the `arguments' array-like object is used as the
params to the driver method.

`typeof arguments[1] === 'object'' is used to allow the parameters to
be passed as a single array, in the same usage style as the driver
methods.