Have native sqlite errors contain stack traces
This ensures that errors thrown from the sqlite driver now
have stack traces and are of an `Error` type.

Thanks to @fresheneesz for initial troubleshooting and initial
code to help fix the issue.

This is a minor level version update as to not break implementations
that may handle errors in their own way prior to this fix.